requirements.txt
django==5.0.6
djangorestframework==3.15.1
django-crum==0.7.9
pillow==10.3.0
drf-yasg
django-cors-headers
djangorestframework-simplejwt
mysqlclient